requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.10.0
click==8.1.3
colorama==0.4.5
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
drf-spectacular==0.24.2
inflection==0.5.1
jsonschema==4.16.0
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
psycopg2==2.9.4
pybase64==1.2.3
PyJWT==2.5.0
pyrsistent==0.18.1
pytz==2022.4
PyYAML==6.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.5
uritemplate==4.1.1
